# docs/mkdocs/api/swagger.yaml
basePath: /
definitions:
...
  api.chainHintJSON:
    properties:
      bech32_prefix:
        type: string
      chain_id:
        type: string
      chain_name:
        type: string
      denom:
        type: string
    type: object
  api.chainRecordJSON:
    properties:
      bech32_prefix:
        type: string
      binary_name:
        type: string
      binary_sha256:
        type: string
      binary_version:
        type: string
      chain_id:
        type: string
      chain_name:
        type: string
      denom:
        type: string
      genesis_time:
        type: string
      gentx_deadline:
        type: string
      max_commission_change_rate:
        type: string
      max_commission_rate:
        type: string
      min_self_delegation:
        type: string
      min_validator_count:
        type: integer
      repo_commit:
        type: string
      repo_url:
        type: string
      total_supply:
        type: string
    type: object
...
  api.committeeMemberJSON:
    properties:
      address:
        type: string
      moniker:
        type: string
      pub_key_b64:
        type: string
    type: object
  api.committeeRequestJSON:
    properties:
      creation_signature:
        type: string
      lead_address:
        type: string
      members:
        items:
          $ref: '#/definitions/api.committeeMemberJSON'
        type: array
      threshold_m:
        type: integer
      total_n:
        type: integer
    type: object
...
  api.createLaunchRequest:
    properties:
      allowlist:
        items:
          type: string
        type: array
      committee:
        $ref: '#/definitions/api.committeeRequestJSON'
      launch_type:
        example: MAINNET
        type: string
      record:
        $ref: '#/definitions/api.chainRecordJSON'
    type: object
...
  api.errorBody:
    properties:
      code:
        type: string
      invariants:
        items:
          $ref: '#/definitions/api.invariantResultJSON'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  api.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/api.errorBody'
    type: object
...
  api.invariantResultJSON:
    properties:
      invariant:
        type: string
      ok:
        type: boolean
      reason:
        type: string
    type: object
...
  api.launchJSON:
    properties:
      created_at:
        type: string
      final_genesis_sha256:
        type: string
      id:
        type: string
      initial_genesis_sha256:
        type: string
      launch_type:
        type: string
      monitor_rpc_url:
        type: string
      record:
        $ref: '#/definitions/api.chainRecordJSON'
      rehearsal_endpoint:
        type: string
      rehearsal_service_pubkey:
        description: RehearsalServicePubKey/RehearsalEndpoint are the bridge fields
          (not secret).
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  api.pageEnvelope-array_api_launchJSON:
    properties:
      items:
        items:
          $ref: '#/definitions/api.launchJSON'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  api.patchLaunchRequest:
    properties:
      allowlist:
        items:
          type: string
        type: array
      binary_sha256:
        type: string
      binary_version:
        type: string
      chain_name:
        type: string
      genesis_time:
        type: string
      min_validator_count:
        type: integer
      monitor_rpc_url:
        type: string
      rehearsal_endpoint:
        type: string
      rehearsal_service_pubkey:
        description: Bridge fields — operational, settable at any status.
        type: string
      repo_commit:
        type: string
      repo_url:
        type: string
      total_supply:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Chain launch coordination server. Coordinators and validators authenticate
    via secp256k1 challenge-response (ADR-036)
  title: ChainCoord API
  version: "1.0"
paths:
...
  /launch:
    post:
      consumes:
      - application/json
      description: Creates a new chain launch. When launch_policy is "restricted",
        the caller must be on the coordinator allowlist.
      parameters:
      - description: Launch definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createLaunchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.launchJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Create a launch
      tags:
      - launches
  /launch/{id}:
    get:
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.launchJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Get a launch
      tags:
      - launches
    patch:
      consumes:
      - application/json
      description: |-
        Partially updates mutable fields on a launch (coordinator only). Most fields require
        DRAFT status; monitor_rpc_url and the rehearsal bridge fields are settable at any status.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Partial update (all fields optional)
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.patchLaunchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.launchJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Update a launch
      tags:
      - launches
...
  /launch/{id}/cancel:
    post:
      description: |-
        Transitions a launch to CANCELED. Only the committee lead may call this.
        No quorum required — cancellation is a single-actor emergency action.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.launchJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "409":
          description: Launch is already in a terminal status
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Cancel a launch
      tags:
      - launches
  /launch/{id}/chain-hint:
    get:
      description: |-
        Returns chain_id, chain_name, bech32_prefix, and denom. Visible only to a launch member
        (committee ∪ allowlist); non-members get 404.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.chainHintJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Chain hint
      tags:
      - launches
...
  /launch/{id}/open-window:
    post:
      description: Transitions a launch from DRAFT to OPEN. Coordinator only.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.launchJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Open the application window
      tags:
      - launches
...
  /launches:
    get:
      description: Returns a paginated list of launches. Visibility is filtered based
        on auth status.
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.pageEnvelope-array_api_launchJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: List launches
      tags:
      - launches
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Session token obtained from POST /auth/verify. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"